fix build errors

+++ src/app/builder/fields/field-action/field-action.component.ts
@@ -1,33 +1,37 @@
 import { Component, OnInit, Input } from "@angular/core";
-import { Fields, FieldType } from '../../model/field';
-import { AppService } from '../../share/Render/app.service';
-import { MatDialog } from '@angular/material/dialog';
-import { ComplexValueComponent } from '../../complex-value/complex-value.component';
+import { Fields, FieldType } from "../../model/field";
+import { AppService } from "../../share/Render/app.service";
+import { MatDialog } from "@angular/material/dialog";
+import { ComplexValueComponent } from "../../complex-value/complex-value.component";
 
 @Component({
   selector: "app-field-action",
   templateUrl: "./field-action.component.html",
-  styleUrls: ["./field-action.component.scss"]
+  styleUrls: ["./field-action.component.scss"],
+  standalone: false,
 })
-
 export class FieldActionComponent implements OnInit {
-
   @Input() field: Fields;
   left: number;
   top: number;
-  constructor(public appService: AppService, public dialog: MatDialog) {
-  }
+  constructor(public appService: AppService, public dialog: MatDialog) {}
 
-  ngOnInit() {
+  ngOnInit() {}
 
-  }
-  handleFieldAction(type, field: Fields) {
-    const containerIndex = this.appService.allContainers.findIndex(x => x.id === field.containerId);
+  handleFieldAction(type: number, field: Fields, event: any) {
+    const containerIndex = this.appService.allContainers.findIndex(
+      (x) => x.id === field.containerId
+    );
     if (containerIndex !== -1) {
-      const index = this.appService.allContainers[containerIndex].fields.findIndex(x => x.id === field.id);
+      const index = this.appService.allContainers[
+        containerIndex
+      ].fields?.findIndex((x) => x.id === field.id);
+      if (index === undefined) {
+        return;
+      }
       switch (type) {
         case 1:
-          this.appService.deleteField(field, index)
+          this.appService.deleteField(field, index);
           break;
         case 2:
           this.appService.copyField(field, index);
@@ -36,28 +40,30 @@ export class FieldActionComponent implements OnInit {
           this.openComplexValueSetting();
           break;
         case 3:
-          this.selectParentContainr(field);
+          this.selectParentContainr(field, event);
           break;
         default:
           break;
       }
     }
   }
-  getType(type) {
-    return FieldType[type].toLowerCase()
+  getType(type: FieldType) {
+    return FieldType[type].toLowerCase();
   }
   openComplexValueSetting() {
-    let setting = { width: '40vw', height: 'auto', data: this.field };
+    let setting = { width: "40vw", height: "auto", data: this.field };
     const dialogRef = this.dialog.open(ComplexValueComponent, setting);
   }
-  selectParentContainr(field: Fields) {
-    const parent = this.appService.allContainers.find(x => x.id === field.containerId);
+
+  selectParentContainr(field: Fields, event: any) {
+    const parent = this.appService.allContainers.find(
+      (x) => x.id === field.containerId
+    );
     if (parent) {
-      this.appService.selectField(parent);
+      this.appService.selectField(parent, event);
     } else {
       //root container
-      this.appService.selectField(this.appService.containers[0]);
+      this.appService.selectField(this.appService.containers[0], event);
     }
-
   }
 }
